big change in code visibility
I thought the main cpp should not have access to any vulkan related code or glfw windows so i made it so that those things are hidden from the rest.

I made a vulkan system class that basically handles the whole vulkan initialization

the vulkan application is the one that completely hides the vulkan system and is the one getting called in the main fil. its also responsible for the window.

the window shouldnt be part of the vulkan system so it also has its own manager class

the main should basically only be able to set some config files and choose the main scene however i will see if i can load objects into a scene if i need to, currently you can only load a singular scene and it also doesnt support setting textures so the obj should have everything implemented correctly