change to sampler allocation
There was an issue where the project didnt run on a specific laptop while in 2 other cases it did. so i made it that the graphics pipeline gets created with right amount of textures that were loaded before its creation.

added a constant to the fragment shader that gets set in the graphics manager when loading the shaders.